daemon: listeners: make systemd listener optional
On non systemd systems, the systemd-based activation is not needed at all,
thus make it build-time optional. It won't be built in if the 'no_systemd'
tag is set.

changes v2: renamed 'nosystemd' tag to 'no_systemd'
changes v3: fixed some obsolete imports
            use new style tags
